Inline the web color space mapping into the WebGL and WebGPU backends and remove the shared helper file.
