+++ src/gpu/opengl/webgl/WebGLDevice.cpp
@@ -18,10 +18,26 @@
 #include "tgfx/gpu/opengl/webgl/WebGLDevice.h"
 #include <emscripten/val.h>
 #include "core/utils/Log.h"
-#include "gpu/WebColorSpace.h"
 #include "gpu/opengl/webgl/WebGLGPU.h"
 
 namespace tgfx {
+
+// Mirrors the WindowColorSpace enum defined in the web binding layer (see web/src/types.ts). The
+// integer values must stay in sync with the TypeScript definition.
+enum class WebNamedColorSpace { None = 0, SRGB = 1, DisplayP3 = 2, Others = 3 };
+
+static WebNamedColorSpace ToWebNamedColorSpace(const std::shared_ptr<ColorSpace>& colorSpace) {
+  if (colorSpace == nullptr) {
+    return WebNamedColorSpace::None;
+  }
+  if (ColorSpace::Equals(colorSpace.get(), ColorSpace::SRGB().get())) {
+    return WebNamedColorSpace::SRGB;
+  }
+  if (ColorSpace::Equals(colorSpace.get(), ColorSpace::DisplayP3().get())) {
+    return WebNamedColorSpace::DisplayP3;
+  }
+  return WebNamedColorSpace::Others;
+}
 
 void* GLDevice::CurrentNativeHandle() {
   return reinterpret_cast<void*>(emscripten_webgl_get_current_context());
