+++ src/gpu/webgpu/WebGPUWindow.cpp
@@ -22,7 +22,6 @@
 #include "WebGPUDrawableProxy.h"
 #include "WebGPUGPU.h"
 #include "core/utils/Log.h"
-#include "gpu/WebColorSpace.h"
 
 #ifdef __EMSCRIPTEN__
 #include <emscripten/html5.h>
@@ -31,6 +30,25 @@
 #endif
 
 namespace tgfx {
+
+#ifdef __EMSCRIPTEN__
+// Mirrors the WindowColorSpace enum defined in the web binding layer (see web/src/types.ts). The
+// integer values must stay in sync with the TypeScript definition.
+enum class WebNamedColorSpace { None = 0, SRGB = 1, DisplayP3 = 2, Others = 3 };
+
+static WebNamedColorSpace ToWebNamedColorSpace(const std::shared_ptr<ColorSpace>& colorSpace) {
+  if (colorSpace == nullptr) {
+    return WebNamedColorSpace::None;
+  }
+  if (ColorSpace::Equals(colorSpace.get(), ColorSpace::SRGB().get())) {
+    return WebNamedColorSpace::SRGB;
+  }
+  if (ColorSpace::Equals(colorSpace.get(), ColorSpace::DisplayP3().get())) {
+    return WebNamedColorSpace::DisplayP3;
+  }
+  return WebNamedColorSpace::Others;
+}
+#endif
 
 std::shared_ptr<WebGPUWindow> WebGPUWindow::MakeFrom(const std::string& canvasSelector,
                                                      std::shared_ptr<WebGPUDevice> device,
